# front-end/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "test task FrontEnd",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/original": {
            "get": {
                "description": "get URL back from short URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "get URL back",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL",
                        "name": "ShortURL",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful returned",
                        "schema": {
                            "$ref": "#/definitions/main.jsonResponse"
                        }
                    }
                }
            }
        },
        "/shorten": {
            "post": {
                "description": "save URL To DB and get short URL back",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "URL shortener",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Long URL",
                        "name": "LongURL",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful shorten",
                        "schema": {
                            "$ref": "#/definitions/main.jsonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.jsonResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}